% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_markers.R
\name{bin_markers}
\alias{bin_markers}
\title{Transform marker expression to categorical variable with fixed number of levels (bins).}
\usage{
bin_markers(marker_expr, num_bins = 8)
}
\description{
Transform marker expression to categorical variable with fixed number of levels (bins).
}
