% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covdm.R
\name{covdm}
\alias{covdm}
\title{Covariance Dirichlet multinomial model using Stan}
\usage{
covdm(df_samples_subset, donors, protein_names, condition, num_boot = 100,
  cell_n_max = 1000, seed = 56026)
}
\description{
Covariance Dirichlet multinomial model using Stan
}
