% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_check.R
\name{cyto_check}
\alias{cyto_check}
\title{Check if input to cytoxxx function have errors}
\usage{
cyto_check(cell_n_subsample, cell_n_min, protein_names)
}
\arguments{
\item{cell_n_subsample}{Subsample samples to have this maximum cell count}

\item{cell_n_min}{A vector of column names of protein to use in the analysis}

\item{protein_names}{A vector of column names of protein to use in the analysis}
}
\description{
Check if input to cytoxxx function have errors
}
