% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoglm.R
\name{cytoglm}
\alias{cytoglm}
\title{Logistic regression with cases bootstrap}
\usage{
cytoglm(
  df_samples_subset,
  protein_names,
  condition,
  group = "donor",
  covariate_names = NULL,
  cell_n_min = Inf,
  cell_n_subsample = 0,
  num_boot = 100,
  num_cores = 1
)
}
\description{
Logistic regression with cases bootstrap
}
