% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoglmm.R
\name{cytoglmm}
\alias{cytoglmm}
\title{Estimate random effects model}
\usage{
cytoglmm(
  df_samples_subset,
  protein_names,
  condition,
  group = "donor",
  covariate_names = NULL,
  cell_n_min = Inf,
  cell_n_subsample = 0,
  num_cores = 1
)
}
\description{
Estimate random effects model
}
