% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytomlogit.R
\name{cytomlogit}
\alias{cytomlogit}
\title{Normal-multinomial regression model using Stan}
\usage{
cytomlogit(df_samples_subset, protein_names, condition, num_boot = 100,
  cell_n_max = 1000, seed = 56026, partition = "normal")
}
\description{
Normal-multinomial regression model using Stan
}
