% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet_multinomial.R
\name{dirichlet_multinomial}
\alias{dirichlet_multinomial}
\title{Latent class regression with mixed effects using Stan.}
\usage{
dirichlet_multinomial(df_samples_subset)
}
\description{
Latent class regression with mixed effects using Stan.
}
