% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate dataset for vignettes and simulation studies}
\usage{
generate_data()
}
\value{
\code{\link[tibble]{tibble}} data frame
}
\description{
Generate dataset for vignettes and simulation studies
}
\examples{
set.seed(23)
df = generate_data()
str(df)
df
}
