% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_status.R
\name{get_status}
\alias{get_status}
\title{Check status of jobs on sherlock cluster}
\usage{
get_status(user = "cseiler", registry = "registry_2017_10_26_113306")
}
\description{
Check status of jobs on sherlock cluster
}
