% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm_ml.R
\name{glmm_ml}
\alias{glmm_ml}
\title{Generalized linear mixed model with maximum likelihood.}
\usage{
glmm_ml(df_samples, protein_names, response, cores = detectCores())
}
\description{
Generalized linear mixed model with maximum likelihood.
}
