% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_unpaired.R
\name{is_unpaired}
\alias{is_unpaired}
\title{Check if samples match or paired on condition}
\usage{
is_unpaired(df_samples_subset, condition, group)
}
\arguments{
\item{df_samples_subset}{Data frame or tibble with proteins counts,
cell condition, and group information}

\item{condition}{The column name of the condition variable}

\item{group}{The column name of the group variable}
}
\value{
A boolean
}
\description{
Check if samples match or paired on condition
}
