% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcr.R
\name{lcr}
\alias{lcr}
\title{Latent class regression with mixed effects using Stan}
\usage{
lcr(df_samples_binned, protein_names, condition, num_latent_classes)
}
\description{
Latent class regression with mixed effects using Stan
}

