% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcr_cv.R
\name{lcr_cv}
\alias{lcr_cv}
\title{Latent class regression with mixed effects using Stan}
\usage{
lcr_cv(df_samples, protein_names, formula, num_latent_classes, num_bins = 8,
  subsample_size = 10000, seed = 1)
}
\description{
Latent class regression with mixed effects using Stan
}
