% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_logit.R
\name{multi_logit}
\alias{multi_logit}
\title{Latent class regression with mixed effects using Stan.}
\usage{
multi_logit(df_samples_subset)
}
\description{
Latent class regression with mixed effects using Stan.
}
