% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_mi.R
\name{pairwise_mi}
\alias{pairwise_mi}
\title{Compute pairwise association matrix wiht mutual information.}
\usage{
pairwise_mi(theta, pi)
}
\description{
Compute pairwise association matrix wiht mutual information.
}
