% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cytoflexmix.R
\name{plot.cytoflexmix}
\alias{plot.cytoflexmix}
\title{Plot all components of mixture regression}
\usage{
\method{plot}{cytoflexmix}(fit, k = NULL, separate = FALSE)
}
\description{
Plot all components of mixture regression
}
