% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cytogroup.R
\name{plot.cytogroup}
\alias{plot.cytogroup}
\title{Plot fixded coefficients of group-specific fixed effects model}
\usage{
plot.cytogroup(fit, order = FALSE, separate = FALSE)
}
\description{
Plot fixded coefficients of group-specific fixed effects model
}
