% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coeff.R
\name{plot_coeff}
\alias{plot_coeff}
\title{Extact coefficient matrix A}
\usage{
plot_coeff(tb, title_str, title_str_right, xlab_str, redline = 0,
  order = FALSE)
}
\description{
Extact coefficient matrix A
}
