% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_donors.R
\name{plot_donors}
\alias{plot_donors}
\title{Plot one density plot faceted by donors.}
\usage{
plot_donors(df_samples, protein_name, histogram = TRUE)
}
\description{
Plot one density plot faceted by donors.
}

