% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_donors_raw.R
\name{plot_donors_raw}
\alias{plot_donors_raw}
\title{Plot row counts faceted by donors}
\usage{
plot_donors_raw(df_samples_subset, condition, protein_name, min = 0,
  max = quantile(df_samples_subset \%>\% pull(protein_name), probs = 0.99))
}
\description{
Plot row counts faceted by donors
}
