% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interaction.R
\name{plot_interaction}
\alias{plot_interaction}
\title{Extact and plot regression coefficients fot total cell count}
\usage{
plot_interaction(fit)
}
\description{
Extact and plot regression coefficients fot total cell count
}
