% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lda.R
\name{plot_lda}
\alias{plot_lda}
\title{LDA on marker expression.}
\usage{
plot_lda(df_samples, protein_names, group, cor_scaling_factor = 1,
  arrow_color = "black", marker_color = "black", marker_size = 5)
}
\description{
LDA on marker expression.
}
