% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_markers.R
\name{plot_markers}
\alias{plot_markers}
\title{Plot all markers for one donor.}
\usage{
plot_markers(df_samples, donor_id, protein_names,
  condition = "condition", density = TRUE)
}
\description{
Plot all markers for one donor.
}
