% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_selection.R
\name{plot_model_selection}
\alias{plot_model_selection}
\title{Plot model selection to choose number optimal number of clusters}
\usage{
plot_model_selection(fit, k = NULL)
}
\description{
Plot model selection to choose number optimal number of clusters
}
