% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Plot protein correlation network derived from coefficient matrix B}
\usage{
plot_network(tb_B, title_str)
}
\description{
Plot protein correlation network derived from coefficient matrix B
}
