% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairwise_mi.R
\name{plot_pairwise_mi}
\alias{plot_pairwise_mi}
\title{Compute pairwise association matrix wiht mutual information.}
\usage{
plot_pairwise_mi(fit, df_samples_binned, protein_names)
}
\description{
Compute pairwise association matrix wiht mutual information.
}

