% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairwise_mi_da.R
\name{plot_pairwise_mi_da}
\alias{plot_pairwise_mi_da}
\title{Plot association matrix difference between stimulation conditions.}
\usage{
plot_pairwise_mi_da(fit, df_samples, protein_names, cores = detectCores())
}
\description{
Plot association matrix difference between stimulation conditions.
}

