% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_total.R
\name{plot_total}
\alias{plot_total}
\title{Extact and plot regression coefficients fot total cell count}
\usage{
plot_total(fit)
}
\description{
Extact and plot regression coefficients fot total cell count
}
