% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.cytomlogit.R
\name{residuals.cytomlogit}
\alias{residuals.cytomlogit}
\title{Extact donor random effect z}
\usage{
\method{residuals}{cytomlogit}(fit, dispersion = TRUE)
}
\description{
Extact donor random effect z
}
