% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cytoglm.R
\name{summary.cytoglm}
\alias{summary.cytoglm}
\title{Extact and calculate p-values of bootstrap GLM fit}
\usage{
\method{summary}{cytoglm}(fit)
}
\description{
Extact and calculate p-values of bootstrap GLM fit
}
