% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoflexmix.R
\name{cytoflexmix}
\alias{cytoflexmix}
\title{Logistic mixture regression}
\usage{
cytoflexmix(df_samples_subset, protein_names, condition, group = "donor",
  cell_n_min = Inf, cell_n_subsample = 0, ks = 1:10, seed = 56026,
  num_cores = 4)
}
\description{
Logistic mixture regression
}
