% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoflexmix.R
\name{cytoflexmix}
\alias{cytoflexmix}
\title{Logistic mixture regression}
\usage{
cytoflexmix(
  df_samples_subset,
  protein_names,
  condition,
  group = "donor",
  cell_n_min = Inf,
  cell_n_subsample = 0,
  ks = seq_len(10),
  num_cores = 1
)
}
\description{
Logistic mixture regression
}
