% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoglm.R
\name{cytoglm}
\alias{cytoglm}
\title{Normal-multinomial regression model using Stan}
\usage{
cytoglm(df_samples_subset, protein_names, condition, unpaired = TRUE,
  cell_n_max = Inf, subsample = FALSE, num_boot = 100, seed = 56026,
  num_cores = 4)
}
\description{
Normal-multinomial regression model using Stan
}
