% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoglmm.R
\name{cytoglmm}
\alias{cytoglmm}
\title{Estimate random effects model}
\usage{
cytoglmm(df_samples_subset, protein_names, condition, cell_n_min = Inf,
  cell_n_subsample = 0, seed = 56026, num_cores = 4)
}
\description{
Estimate random effects model
}
