% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytogroup.R
\name{cytogroup}
\alias{cytogroup}
\title{Group-specific fixed effects model}
\usage{
cytogroup(df_samples_subset, protein_names, condition, cell_n_min = Inf,
  cell_n_subsample = 0, seed = 56026)
}
\description{
Group-specific fixed effects model
}
