% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytostab.R
\name{cytostab}
\alias{cytostab}
\title{Evaluate parameter stability with respect to gating sheme.}
\usage{
cytostab(
  df_samples_subset,
  protein_names,
  condition,
  group = "donor",
  cell_n_min = Inf,
  cell_n_subsample = 0
)
}
\description{
Evaluate parameter stability with respect to gating sheme.
}
