% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm_moment.R
\name{glmm_moment}
\alias{glmm_moment}
\title{Generalized linear mixed model with maximum likelihood.}
\usage{
glmm_moment(
  df_samples,
  protein_names,
  response,
  group = "donor",
  covariate_names = NULL,
  num_cores = 1
)
}
\description{
Generalized linear mixed model with maximum likelihood.
}
