% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_unpaired.R
\name{is_unpaired}
\alias{is_unpaired}
\title{Check if samples match or paired on condition}
\usage{
is_unpaired(df_samples_subset, condition, group)
}
\description{
Check if samples match or paired on condition
}
