% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cytoglmm.R
\name{plot.cytoglmm}
\alias{plot.cytoglmm}
\title{Plot fixded coefficients of random effects model}
\usage{
\method{plot}{cytoglmm}(fit, order = FALSE, separate = FALSE)
}
\description{
Plot fixded coefficients of random effects model
}
