% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coeff.R
\name{plot_coeff}
\alias{plot_coeff}
\title{Helper function to plot regression coeffcient}
\usage{
plot_coeff(tb, title_str, title_str_right, xlab_str, redline = 0,
  order = FALSE, separate = FALSE)
}
\description{
Helper function to plot regression coeffcient
}
