% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coeff.R
\name{plot_coeff}
\alias{plot_coeff}
\title{Helper function to plot regression coeffcient}
\usage{
plot_coeff(
  tb,
  title_str,
  title_str_right,
  xlab_str,
  redline = 0,
  order = FALSE,
  separate = FALSE
)
}
\arguments{
\item{tb}{A data frame}

\item{title_str}{Title string for summary plot}

\item{title_str_right}{Title for bootstrap sample plot}

\item{xlab_str}{Label on x-axis}

\item{redline}{Point on x-axis to draw the red line}

\item{order}{Order the markers according to the mangintute of the coefficients}

\item{separate}{Plot both summary and bootstrap samples}
}
\value{
\code{\link[ggplot2]{ggplot2}} object or list of two objects if separate is true
}
\description{
Helper function to plot regression coeffcient
}
