% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Heatmap of median marker expression.}
\usage{
plot_heatmap(df_samples, sample_info_names, protein_names, arrange_by, rownames)
}
\description{
Heatmap of median marker expression.
}
