% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairwise_mi.R
\name{plot_pairwise_mi}
\alias{plot_pairwise_mi}
\title{Plot two association matrices for both stimulation conditions.}
\usage{
plot_pairwise_mi(fit, df_samples, protein_names)
}
\description{
Plot two association matrices for both stimulation conditions.
}
