% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prcomp.R
\name{plot_prcomp}
\alias{plot_prcomp}
\title{Plot PCA of subsampled data using ggplot.}
\usage{
plot_prcomp(
  df_samples,
  protein_names,
  color_var = "treatment",
  subsample_size = 10000,
  repel = TRUE
)
}
\description{
Plot PCA of subsampled data using ggplot.
}
