% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_samples.R
\name{remove_samples}
\alias{remove_samples}
\title{Remove samples based on low cell counts}
\usage{
remove_samples(df_samples_subset, condition, group, unpaired, cell_n_min)
}
\arguments{
\item{df_samples_subset}{Data frame or tibble with proteins counts,
cell condition, and group information}

\item{condition}{The column name of the condition variable}

\item{group}{The column name of the group variable}

\item{unpaired}{true if unpaired samples were provided as input}

\item{cell_n_min}{Remove samples that are below this cell counts threshold}
}
\description{
Remove samples based on low cell counts
}
