% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cytoglmm.R
\name{summary.cytoglmm}
\alias{summary.cytoglmm}
\title{Extact and calculate p-values of GLMM fit}
\usage{
\method{summary}{cytoglmm}(fit, method = "BH")
}
\description{
Extact and calculate p-values of GLMM fit
}
